# services/gateway/server/api/gateway/rest/handers/users/data/data.go
package data

import (
	pb "gateway/server/api/gateway/grpc/gen/users"
	userclient "gateway/server/api/gateway/grpc/user_client"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/render"
	"github.com/google/uuid"
	"go.uber.org/zap"
)

type UserResponseDTO struct {
	pb.UserResponse
}

// @Summary Обновление статуса заказа
// @Description Возвращает объект заказа
// @Tags users
// @Accept json
// @Produce json
// @Param uuid path string true "uuid пользователя"
// @Param Authorization header string true "Bearer token"
// @Success 200 {object} UserResponseDTO
// @Router /api/users/{uuid} [get]
func GetUser(log *zap.Logger, client *userclient.UserGRPCClient) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		userUuid := chi.URLParam(r, "uuid")
		_, err := uuid.Parse(userUuid)
		if err != nil {
			http.Error(w, "invalid uuid", http.StatusBadRequest)
		}
		resp, err := client.GetUser(userUuid)
		if err != nil {
			http.Error(w, "failed to get users", http.StatusInternalServerError)
			log.Error(err.Error())
			return
		}

		render.JSON(w, r, resp)
	}
}